test of GearVR support for VRTK
**made with Unity 5.5.2**

In Assets\VRTK_GearVR_Test here are 3 scenes to be tested with GearVR:
* object touching and grabbing (005_Controller_BasicObjectGrabbing_GearVR)
* interaction with Unity UI (034_Controls_InteractingWithUnityUI_GearVR)
* teleport (044_CameraRig_RestrictedTeleportZones_GearVR)

The examples were tested after changing `[CameraRig]` and `[VRTK]` in
order to support GearVR. The script `EditorLikeCameraController` is
attached to the camera to move/rotate it only in Unity Editor play mode.
In this way the examples can be tested also with mouse and keyboard
(the controls are almost the same as in Unity Scene Editor).

**Warning!** You must copy special OSIG file additions (`oculussig_*`)
 to `Assets/Plugins/Android/Assets/`